+++ packages/core/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact core module, newest changes on the top.
 
+## [4.0.1] - 2021-05-21
+
+### Fixed
+
+- `core/platform` to detect chromium version for webOS
+
 ## [4.0.0] - 2021-03-26
 
 No significant changes.